refactor: move all optimizations and pre/post fork hooks to separate file
Now they will truly execute before/after fork = :pinch: few bytes saved!
